// src/Actions/type.ts
export enum TICKER {
    UPATE_BID = 'UPATE_BID',
    UPATE_BID_SIZE = 'UPATE_BID_SIZE',
    UPATE_ASK = 'UPATE_ASK',
    UPATE_ASK_SIZE = 'UPATE_ASK_SIZE',
    UPATE_DAILY_CHANGE = 'UPATE_DAILY_CHANGE',
    UPATE_DAILY_CHANGE_RELATIVE = 'UPATE_DAILY_CHANGE_RELATIVE',
    UPATE_LAST_PRICE = 'UPATE_LAST_PRICE',
    UPATE_VOLUME = 'UPATE_VOLUME',
    UPATE_HIGH = 'UPATE_HIGH',
    UPATE_LOW = 'UPATE_LOW'
}